+++ account/config_test.go
@@ -16,9 +16,9 @@ import (
 	"strings"
 	"testing"
 
-	triton "github.com/joyent/triton-go"
 	"github.com/joyent/triton-go/account"
 	"github.com/joyent/triton-go/testutils"
+	"github.com/joyent/triton-go/triton"
 	"github.com/pkg/errors"
 )
