+++ compute/datacenters_test.go
@@ -19,9 +19,9 @@ import (
 	"testing"
 
 	"github.com/abdullin/seq"
-	triton "github.com/joyent/triton-go"
 	"github.com/joyent/triton-go/compute"
 	"github.com/joyent/triton-go/testutils"
+	"github.com/joyent/triton-go/triton"
 )
 
 const dataCenterName = "us-east-1"
