+++ compute/instances_test.go
@@ -25,10 +25,10 @@ import (
 
 	"path"
 
-	triton "github.com/joyent/triton-go"
 	"github.com/joyent/triton-go/compute"
 	"github.com/joyent/triton-go/network"
 	"github.com/joyent/triton-go/testutils"
+	"github.com/joyent/triton-go/triton"
 )
 
 var (
