+++ compute/packages_test.go
@@ -20,9 +20,9 @@ import (
 	"testing"
 
 	"github.com/abdullin/seq"
-	triton "github.com/joyent/triton-go"
 	"github.com/joyent/triton-go/compute"
 	"github.com/joyent/triton-go/testutils"
+	"github.com/joyent/triton-go/triton"
 )
 
 var (
