+++ identity/client.go
@@ -9,8 +9,8 @@
 package identity
 
 import (
-	triton "github.com/joyent/triton-go"
 	"github.com/joyent/triton-go/client"
+	"github.com/joyent/triton-go/triton"
 )
 
 type IdentityClient struct {
