+++ identity/users_test.go
@@ -19,9 +19,9 @@ import (
 	"strings"
 	"testing"
 
-	triton "github.com/joyent/triton-go"
 	"github.com/joyent/triton-go/identity"
 	"github.com/joyent/triton-go/testutils"
+	"github.com/joyent/triton-go/triton"
 )
 
 const accountUrl = "testing"
